% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assignConsequence}
\alias{assignConsequence}
\title{assignConsequence function
Assign a main consequence from the consequences returned by VEP}
\usage{
assignConsequence(consequence)
}
\arguments{
\item{consequence}{consequence from VEP}
}
\value{
main consequence
}
\description{
assignConsequence function
Assign a main consequence from the consequences returned by VEP
}
