% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assignPopulationInfo}
\alias{assignPopulationInfo}
\title{assignPopulationInfo function
Search for the variant in population datasets and extract required information}
\usage{
assignPopulationInfo(variant_info, gnomad4, kova_tommo, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{gnomad4}{gnomad population dataset}

\item{kova_tommo}{kova and tommo east asian population dataset}

\item{config_values}{configurations}
}
\value{
variant_info named list with all population information added
}
\description{
assignPopulationInfo function
Search for the variant in population datasets and extract required information
}
