% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assignProximateCurations}
\alias{assignProximateCurations}
\title{assignProximateCurations function
Search for the variants in proximity of this variant in curated dataset}
\usage{
assignProximateCurations(variant_info, bulk_curations, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{bulk_curations}{bulk curations dataset for PM5 and PS1}

\item{config_values}{configurations}
}
\value{
variant_info named list with nearby bariant info added
}
\description{
assignProximateCurations function
Search for the variants in proximity of this variant in curated dataset
}
