% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_updated_Pr_multi}
\alias{calculate_updated_Pr_multi}
\title{calculate_updated_Pr_multi}
\usage{
calculate_updated_Pr_multi(
  n,
  k,
  l,
  m,
  p = c(0.58, 0.1, 0.007),
  q = c(0.002, 0.002, 0.0015),
  pr_A = 0.1
)
}
\arguments{
\item{n}{is the denominator, no. of germline variant observed}

\item{k}{is the number of times a single recurrent somatic hotspot is seen}

\item{l}{is the number of times a single non-recurrent somatic variant is seen}

\item{m}{is the number of times a multi-somatic hit is seen}

\item{p}{is the expected probabilities of observing k, l, m (out of n) in pathogenic variants}

\item{q}{is the expected probabilities of observing k, l, m (out of n) in cases without a germline variant as a surrogate of it being benign}

\item{pr_A}{posterior probability of germline variant being pathogenic/deleterious}
}
\value{
calculated posterior probability
}
\description{
calculate_updated_Pr_multi
}
