% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculateOddsValues}
\alias{calculateOddsValues}
\title{calculateOddsValues function
Calculate ODDs ratios for PS4 and PP4}
\usage{
calculateOddsValues(variant_info, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{config_values}{configurations}
}
\value{
variant_info named list with odds ratios
}
\description{
calculateOddsValues function
Calculate ODDs ratios for PS4 and PP4
}
