% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calulateOddsRatio}
\alias{calulateOddsRatio}
\title{calulateOddsRatio function
Calculate Odds Ratio and CIs}
\usage{
calulateOddsRatio(variant_info, ethnicity)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{ethnicity}{ethnicity choice}
}
\value{
odds ratio, lower bound and upper bound of 95 percent confidence intervals
}
\description{
calulateOddsRatio function
Calculate Odds Ratio and CIs
}
