% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{canonicalSpliceSite}
\alias{canonicalSpliceSite}
\title{canonicalSpliceSite function
Check whether a canonical splice site variant}
\usage{
canonicalSpliceSite(hgvsc)
}
\arguments{
\item{hgvsc}{hgvsc}
}
\value{
canonical splice site variant or not
}
\description{
canonicalSpliceSite function
Check whether a canonical splice site variant
}
