% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkIntronicSite}
\alias{checkIntronicSite}
\title{checkIntronicSite function
Check whether this intronic variant is beyond +7 or -21}
\usage{
checkIntronicSite(hgvsc)
}
\arguments{
\item{hgvsc}{hgvsc}
}
\value{
intronic variant is beyond +7 or -21
}
\description{
checkIntronicSite function
Check whether this intronic variant is beyond +7 or -21
}
