% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkSynonymousSite}
\alias{checkSynonymousSite}
\title{checkSynonymousSite function
Check whether this synonymous variant is in last 3 bases of an exon or 1 st base of an exon}
\usage{
checkSynonymousSite(coding_start, exon)
}
\arguments{
\item{coding_start}{coding start}

\item{exon}{exon}
}
\value{
T if synonymous variant is in last 3 bases of an exon or 1 st base of an exon
}
\description{
checkSynonymousSite function
Check whether this synonymous variant is in last 3 bases of an exon or 1 st base of an exon
}
