% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateVariant.R
\name{curateVariant}
\alias{curateVariant}
\title{Method for curating a variant given in variant_info}
\usage{
curateVariant(variant_info, gnomad4, kova_tommo, bulk_curations, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{gnomad4}{gnomad population dataset}

\item{kova_tommo}{kova and tommo east asian population dataset}

\item{bulk_curations}{bulk curations dataset for PM5 and PS1}

\item{config_values}{configurations}
}
\value{
list with curation information
\itemize{
  \item decision: Final curation
  \item rule_curations: Individual rule curations in a named list
}
}
\description{
Method for curating a variant given in variant_info
}
