% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluate_splice}
\alias{evaluate_splice}
\title{evaluate_splice function
evaluate splice site or intronic variant}
\usage{
evaluate_splice(hgvsc, variant_info, eval_PVS1, eval_PP3)
}
\arguments{
\item{hgvsc}{hgvsc}

\item{variant_info}{named list with all the information required for curation}

\item{eval_PVS1}{PVS1 decision}

\item{eval_PP3}{PP3 decision}
}
\value{
curation
}
\description{
evaluate_splice function
evaluate splice site or intronic variant
}
