% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluateBA1}
\alias{evaluateBA1}
\title{evaluateBA1 function
BA1 curation
Allele frequency is > 5 percent in Exome Sequencing Project, 1000 Genomes Project, or Exome Aggregation Consortium.}
\usage{
evaluateBA1(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluateBA1 function
BA1 curation
Allele frequency is > 5 percent in Exome Sequencing Project, 1000 Genomes Project, or Exome Aggregation Consortium.
}
