% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluateBP4}
\alias{evaluateBP4}
\title{evaluateBP4 function
BP4 curation
Multiple lines of computational evidence suggest no impact on gene or gene product.}
\usage{
evaluateBP4(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluateBP4 function
BP4 curation
Multiple lines of computational evidence suggest no impact on gene or gene product.
}
