% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluateBP7}
\alias{evaluateBP7}
\title{evaluateBP7 function
BP7 curation
A synonymous (silent) variant for which splicing prediction algorithms predict no impact to the splice consensus
     sequence nor the creation of a new splice site AND the nucleotide is not highly conserved.}
\usage{
evaluateBP7(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluateBP7 function
BP7 curation
A synonymous (silent) variant for which splicing prediction algorithms predict no impact to the splice consensus
     sequence nor the creation of a new splice site AND the nucleotide is not highly conserved.
}
