% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluateBS1}
\alias{evaluateBS1}
\title{evaluateBS1 function
BS1 curation
Allele frequency is greater than expected for disorder.}
\usage{
evaluateBS1(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluateBS1 function
BS1 curation
Allele frequency is greater than expected for disorder.
}
