% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePM2}
\alias{evaluatePM2}
\title{evaluatePM2 function
PM2 curation
GroupMax FAF applied for now
Possibly later change to not applicable VS absent/rare in population databases}
\usage{
evaluatePM2(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluatePM2 function
PM2 curation
GroupMax FAF applied for now
Possibly later change to not applicable VS absent/rare in population databases
}
