% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePM4}
\alias{evaluatePM4}
\title{evaluatePM4 function
PM4 curation
Protein length changes due to in-frame deletions/ insertion in a non-repeat region or
  stop loss variants}
\usage{
evaluatePM4(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluatePM4 function
PM4 curation
Protein length changes due to in-frame deletions/ insertion in a non-repeat region or
  stop loss variants
}
