% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePM5}
\alias{evaluatePM5}
\title{evaluatePM5 function
PM5 curation
Missense change at an amino acid residue where a different missense change which has
  been determined to be pathogenic before.
Nonsense/frameshift variants throughout the gene in transcript NM_016222.4}
\usage{
evaluatePM5(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluatePM5 function
PM5 curation
Missense change at an amino acid residue where a different missense change which has
  been determined to be pathogenic before.
Nonsense/frameshift variants throughout the gene in transcript NM_016222.4
}
