% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePP1}
\alias{evaluatePP1}
\title{evaluatePP1 function
PP1 curation
Co-segregation with disease in multiple affected family members}
\usage{
evaluatePP1(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluatePP1 function
PP1 curation
Co-segregation with disease in multiple affected family members
}
