% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePP3}
\alias{evaluatePP3}
\title{evaluatePP3 function
PP3 curation
Multiple lines of computational evidence support a deleterious effect on the gene or gene product}
\usage{
evaluatePP3(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluatePP3 function
PP3 curation
Multiple lines of computational evidence support a deleterious effect on the gene or gene product
}
