% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePP4}
\alias{evaluatePP4}
\title{evaluatePP4 function
PP4 curation
Patient's phenotype or family history is highly specific for a disease with a single
 genetic etiology}
\usage{
evaluatePP4(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluatePP4 function
PP4 curation
Patient's phenotype or family history is highly specific for a disease with a single
 genetic etiology
}
