% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePS1}
\alias{evaluatePS1}
\title{evaluatePS1 function
PS1 curation
Same amino acid change as a previously established P/LP variant regardless of the nucleotide change (must be reviewed by MM-VCEP)
For missense variants, RNA data or splicing predictor show no imapct on splicing
Not applicable for splice site variants, except for the
  canonical (+/-2) (DO NOT apply for +2G>C variants)
  U2 donor motif (last 3 bases of the exon and 6 nucleotides of the intron),
  or the U2 acceptor motif (20 nucleotides of the intron and 1st base of the exon).
Splicing predictions for the variant being evaluated and the known P/LP should match.}
\usage{
evaluatePS1(variant_info, curation)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}
}
\value{
named list with curations
}
\description{
evaluatePS1 function
PS1 curation
Same amino acid change as a previously established P/LP variant regardless of the nucleotide change (must be reviewed by MM-VCEP)
For missense variants, RNA data or splicing predictor show no imapct on splicing
Not applicable for splice site variants, except for the
  canonical (+/-2) (DO NOT apply for +2G>C variants)
  U2 donor motif (last 3 bases of the exon and 6 nucleotides of the intron),
  or the U2 acceptor motif (20 nucleotides of the intron and 1st base of the exon).
Splicing predictions for the variant being evaluated and the known P/LP should match.
}
