% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePS4}
\alias{evaluatePS4}
\title{evaluatePS4 function
PS4 curation
Prevalence of the variant in affected individuals is significantly increased compared
with the prevalence in controls}
\usage{
evaluatePS4(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluatePS4 function
PS4 curation
Prevalence of the variant in affected individuals is significantly increased compared
with the prevalence in controls
}
