% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluatePVS1}
\alias{evaluatePVS1}
\title{evaluatePVS1 function
PVS1 curation
Null variant in a gene where loss of function is a known mechanism of disease}
\usage{
evaluatePVS1(variant_info, curation, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{curation}{named list with curation results}

\item{config_values}{configurations}
}
\value{
named list with curations
}
\description{
evaluatePVS1 function
PVS1 curation
Null variant in a gene where loss of function is a known mechanism of disease
}
