% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_max_spliceAI}
\alias{extract_max_spliceAI}
\title{extract_max_spliceAI function
Extract max spliceAI score}
\usage{
extract_max_spliceAI(splice)
}
\arguments{
\item{splice}{dataframe with splice AI scores}
}
\value{
max spliceAI score
}
\description{
extract_max_spliceAI function
Extract max spliceAI score
}
