% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getTerminalPos}
\alias{getTerminalPos}
\title{getTerminalPos function
Get terminal position of a hgvsp and convert to coding position
If not found check hgvsc}
\usage{
getTerminalPos(hgvsp, hgvsc, exon, config_values)
}
\arguments{
\item{hgvsp}{hgvsp}

\item{hgvsc}{hgvsc}

\item{exon}{exon}

\item{config_values}{configurations}
}
\value{
terminal position
}
\description{
getTerminalPos function
Get terminal position of a hgvsp and convert to coding position
If not found check hgvsc
}
