% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{handle_canonical}
\alias{handle_canonical}
\title{handle_canonical function}
\usage{
handle_canonical(eval_nearby_canonical, eval_nearby_motif, eval_PVS1)
}
\arguments{
\item{eval_nearby_canonical}{evaluate nearby canonical variants}

\item{eval_nearby_motif}{evaluate nearby motif}

\item{eval_PVS1}{PVS1 decision}
}
\value{
list with curation and curation_log
}
\description{
handle_canonical function
}
