% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{handle_motif}
\alias{handle_motif}
\title{handle_motif function}
\usage{
handle_motif(
  eval_same_nucleotide,
  eval_nearby_canonical,
  eval_nearby_motif,
  eval_PP3
)
}
\arguments{
\item{eval_same_nucleotide}{evaluate same nucleotide variants}

\item{eval_nearby_canonical}{evaluate nearby canonical variants}

\item{eval_nearby_motif}{evaluate nearby motif}

\item{eval_PP3}{PP3 decision}
}
\value{
curation
}
\description{
handle_motif function
}
