% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nearby_canonical}
\alias{nearby_canonical}
\title{nearby_canonical function
Get the nearby canonical nucleotide positions}
\usage{
nearby_canonical(hgvsc)
}
\arguments{
\item{hgvsc}{hgvsc}
}
\value{
nearby canonical nucleotide positions
}
\description{
nearby_canonical function
Get the nearby canonical nucleotide positions
}
