% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nearby_motif}
\alias{nearby_motif}
\title{nearby_motif function
Get the nearby U2 motif nucleotide positions}
\usage{
nearby_motif(hgvsc)
}
\arguments{
\item{hgvsc}{hgvsc}
}
\value{
nearby U2 motif nucleotide positions
}
\description{
nearby_motif function
Get the nearby U2 motif nucleotide positions
}
