% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_hgvsp}
\alias{normalize_hgvsp}
\title{normalize_hgvsp function
normalize hgvsp for partial matches}
\usage{
normalize_hgvsp(hgvsp)
}
\arguments{
\item{hgvsp}{original hgvsp}
}
\value{
normalised hgvsp
}
\description{
normalize_hgvsp function
normalize hgvsp for partial matches
}
