% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pathogenicityDecision}
\alias{pathogenicityDecision}
\title{pathogenicityDecision function
Pathogenicity decision rules}
\usage{
pathogenicityDecision(lst_curation)
}
\arguments{
\item{lst_curation}{list of curations, one list item per rule}
}
\value{
decision
}
\description{
pathogenicityDecision function
Pathogenicity decision rules
}
