% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{runCurations}
\alias{runCurations}
\title{Main method for curating a ddx41 variant, this method expects the variant_info named list
to have information in a specific format}
\usage{
runCurations(variant_info, config_values)
}
\arguments{
\item{variant_info}{named list with all the information required for curation}

\item{config_values}{configurations}
}
\value{
variant_info with curation results
}
\description{
Main method for curating a ddx41 variant, this method expects the variant_info named list
to have information in a specific format
}
