% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spliceJunction}
\alias{spliceJunction}
\title{spliceJunction function
Check whether the variant is in a splice junction and returns if "Canonical", "Motif" or "Exception"}
\usage{
spliceJunction(hgvsc)
}
\arguments{
\item{hgvsc}{hgvsc}
}
\value{
splice junction info
}
\description{
spliceJunction function
Check whether the variant is in a splice junction and returns if "Canonical", "Motif" or "Exception"
}
