% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spliceJunctionNotSupported}
\alias{spliceJunctionNotSupported}
\title{spliceJunctionNotSupported function
Check whether the variant is in a splice junction +/- 1 or 2 only but del, ins, or delins}
\usage{
spliceJunctionNotSupported(hgvsc)
}
\arguments{
\item{hgvsc}{hgvsc}
}
\value{
vartiant in a splice junction or not
}
\description{
spliceJunctionNotSupported function
Check whether the variant is in a splice junction +/- 1 or 2 only but del, ins, or delins
}
